+++ changelog.md
@@ -3,7 +3,7 @@
 Good bye old interface.
 
 New features:
-- Integrate Youtube's block/report menu in current session of embedded ytcFilter so as to fight against spam even for those without the extension. Note that due to the new fetch interceptor, the menu integration may take a minute to appear. It may also fail due to Youtube poor performance
+- Integrate Youtube's block/report menu in current session of embedded ytcFilter so as to fight against spam even for those without the extension. Note that due to the new fetch interceptor, the menu integration may take a minute to appear. It may also fail due to Youtube poor performance. Also, due to Youtube's limitation and performance, the menu can't be used if the message is no longer in the chat, so it's not permanent and may also be missing.
 - Drag resize of embedded ytcFIlter
 - Configurable automatic storage management
 - Alert every minute if less than 5% of storage is available (Firefox is not supported)